requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.21
asyncpg>=0.29.0
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
cryptography>=41.0.0
httpx>=0.25.0
apscheduler>=3.10.0
orjson>=3.9.0
